+++ test/2-timers.js
@@ -11,10 +11,12 @@ const queue = new Queue(2, {
 queue
     .push(timeout, 200)
     .push(timeout, 900)
-    .push(timeout, 1100)
+    .push(timeout, 15000)
     .push(timeout, 300)
+    .timeout(1000)
     .success((data, args) => {
         if (data) console.log(data);
 
         console.log(`Timer with time ${args} expired`)
     })
+    .complete(() => console.log("Execution finished"))
